refactor to ParallelBackend
