+++ pycaret/parallel_backends/__init__.py
@@ -1,0 +1,7 @@
+from typing import Any
+from .parallel_backend import ParallelBackend, NoDisplay
+
+try:
+    from .fugue_backend import FugueBackend
+except ImportError:
+    pass
