+++ requirements-optional.txt
@@ -17,6 +17,7 @@ fairlearn # for check_fairness function
 fastapi # for web api
 uvicorn # for web api
 gradio # for gradio app
+fugue>=0.6.5 # for distributed
 
 # for cloud deployment
 boto3
